Всё без тестов c разбивкой на файлы new4

+++ sources/parser-lab-1.cpp
@@ -137,7 +137,7 @@ void print(const std::vector<Student>& students, int len[4]) {
 int main() {
   const ::string File =
       "/home/FORTYSS/Рабочий стол/1/lab-parser-01/sources/Student.json";
-  int len[4] = {11, 3, 3, 11};
+  size_t len[4] = {11, 3, 3, 11};
   std::vector<Student> students = parser(File, len) ;
   print(students, len);
   return 0;
